Make extensions to sealed types final
This enables us to use the `LeakingSealed` wart after #23 is merged.
Given we wrote this code (and not the students) we should make sure
we use the best practices we can.

From the Wartremover documentation for [LeakingSealed](http://www.wartremover.org/doc/warts.html#leakingsealed):

> Descendants of a sealed type must be final or sealed.
> Otherwise this type can be extended in another file through its
> descendant

There's a similar wart called `FinalCaseClass` that we have disabled
in #23, because it would be annoying for the students to add `final` in
front of every case class they write.